move diff cmd tests to own module
